file_data: move tests into tests/
